# docs/full/internal_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for flight data ingest, job status, route performance, and carrier performance. Protected routes require an API key (`Authorization: Bearer \u003ckey\u003e` or `X-API-Key`). `/health` and `/ready` are unauthenticated. Rate limits are shared in Postgres across API replicas; 429 responses include `Retry-After`, `X-RateLimit-Limit`, `X-RateLimit-Remaining`, and `X-RateLimit-Reset`.",
        "title": "flight-tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/carriers/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Aggregated on-time, delay, cancellation, and diversion stats for a marketing carrier, plus best and worst routes and airports (on-time rate, min 30 flights, top/bottom 5). Carrier is a 2-letter marketing IATA code. Dates are optional together and default to the trailing 90 days ending at the carrier's latest flight date (max span 366 days). State is a 2-letter code matching origin or dest.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carriers",
                    "external"
                ],
                "summary": "Carrier performance stats",
                "parameters": [
                    {
                        "maxLength": 2,
                        "minLength": 2,
                        "type": "string",
                        "description": "Marketing carrier code",
                        "name": "carrier",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Range start (YYYY-MM-DD); required if end_date is set",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Range end (YYYY-MM-DD), on or after start_date; required if start_date is set",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "maxLength": 2,
                        "minLength": 2,
                        "type": "string",
                        "description": "Origin or dest state code",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CarrierStats"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/keys": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all API keys, including revoked keys. Hashes and plaintext secrets are never included.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keys",
                    "internal"
                ],
                "summary": "List API keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.APIKeyResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a hashed API key. The plaintext key is returned once; subsequent authentication uses Authorization: Bearer or X-API-Key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keys",
                    "internal"
                ],
                "summary": "Create API key",
                "parameters": [
                    {
                        "description": "Role and optional name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreatedAPIKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/keys/{id}/revoke": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Revokes an API key. Already-revoked keys return the current metadata.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keys",
                    "internal"
                ],
                "summary": "Revoke API key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.APIKeyResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prefix": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/model.APIKeyRole"
                }
            }
        },
        "handlers.CreatedAPIKeyResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prefix": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/model.APIKeyRole"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.APIKeyRole": {
            "type": "string",
            "enum": [
                "consumer",
                "subscriber",
                "admin"
            ],
            "x-enum-varnames": [
                "APIKeyRoleConsumer",
                "APIKeyRoleSubscriber",
                "APIKeyRoleAdmin"
            ]
        },
...
        "model.CarrierAirportStat": {
            "type": "object",
            "properties": {
                "airport": {
                    "type": "string"
                },
                "cancellation_rate": {
                    "type": "number"
                },
                "cancelled": {
                    "type": "integer"
                },
                "delay_rate": {
                    "type": "number"
                },
                "delayed": {
                    "type": "integer"
                },
                "diversion_rate": {
                    "type": "number"
                },
                "diverted": {
                    "type": "integer"
                },
                "flights": {
                    "type": "integer"
                },
                "on_time": {
                    "type": "integer"
                },
                "on_time_rate": {
                    "type": "number"
                }
            }
        },
        "model.CarrierRouteStat": {
            "type": "object",
            "properties": {
                "cancellation_rate": {
                    "type": "number"
                },
                "cancelled": {
                    "type": "integer"
                },
                "delay_rate": {
                    "type": "number"
                },
                "delayed": {
                    "type": "integer"
                },
                "dest": {
                    "type": "string"
                },
                "diversion_rate": {
                    "type": "number"
                },
                "diverted": {
                    "type": "integer"
                },
                "flights": {
                    "type": "integer"
                },
                "on_time": {
                    "type": "integer"
                },
                "on_time_rate": {
                    "type": "number"
                },
                "origin": {
                    "type": "string"
                }
            }
        },
        "model.CarrierStats": {
            "type": "object",
            "properties": {
                "avg_arrival_delay_minutes": {
                    "type": "number"
                },
                "avg_arrival_delay_when_delayed": {
                    "type": "number"
                },
                "avg_departure_delay_minutes": {
                    "type": "number"
                },
                "avg_departure_delay_when_delayed": {
                    "type": "number"
                },
                "best_airports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CarrierAirportStat"
                    }
                },
                "best_routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CarrierRouteStat"
                    }
                },
                "cancellation_rate": {
                    "type": "number"
                },
                "cancelled": {
                    "type": "integer"
                },
                "carrier": {
                    "type": "string"
                },
                "delay_causes_avg_minutes": {
                    "$ref": "#/definitions/model.DelayCausesAvgMinutes"
                },
                "delay_causes_share": {
                    "$ref": "#/definitions/model.DelayCausesShare"
                },
                "delay_rate": {
                    "type": "number"
                },
                "delayed": {
                    "type": "integer"
                },
                "diversion_rate": {
                    "type": "number"
                },
                "diverted": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "filters": {
                    "$ref": "#/definitions/model.CarrierStatsFilters"
                },
                "flights": {
                    "type": "integer"
                },
                "median_arrival_delay_minutes": {
                    "type": "number"
                },
                "median_arrival_delay_when_delayed": {
                    "type": "number"
                },
                "on_time": {
                    "type": "integer"
                },
                "on_time_rate": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "worst_airports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CarrierAirportStat"
                    }
                },
                "worst_routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CarrierRouteStat"
                    }
                }
            }
        },
        "model.CarrierStatsFilters": {
            "type": "object",
            "properties": {
                "state": {
                    "type": "string"
                }
            }
        },
        "model.CreateAPIKeyRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "model.DelayCausesAvgMinutes": {
            "type": "object",
            "properties": {
                "carrier": {
                    "type": "number"
                },
                "late_aircraft": {
                    "type": "number"
                },
                "nas": {
                    "type": "number"
                },
                "security": {
                    "type": "number"
                },
                "weather": {
                    "type": "number"
                }
            }
        },
        "model.DelayCausesShare": {
            "type": "object",
            "properties": {
                "carrier": {
                    "type": "number"
                },
                "late_aircraft": {
                    "type": "number"
                },
                "nas": {
                    "type": "number"
                },
                "security": {
                    "type": "number"
                },
                "unattributed": {
                    "type": "number"
                },
                "weather": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key as `Bearer \u003ckey\u003e`. `X-API-Key` is also accepted.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}